moved Sauce Labs specific capabilities to a new class
