+++ thucydides-core/src/main/java/net/thucydides/core/webdriver/WebDriverFactory.java
@@ -12,6 +12,7 @@ import net.thucydides.core.steps.FilePathParser;
 import net.thucydides.core.util.EnvironmentVariables;
 import net.thucydides.core.util.NameConverter;
 import net.thucydides.core.webdriver.capabilities.BrowserStackRemoteDriverCapabilities;
+import net.thucydides.core.webdriver.capabilities.SauceRemoteDriverCapabilities;
 import net.thucydides.core.webdriver.chrome.OptionsSplitter;
 import net.thucydides.core.webdriver.firefox.FirefoxProfileEnhancer;
 import net.thucydides.core.webdriver.phantomjs.PhantomJSCapabilityEnhancer;
@@ -77,6 +78,7 @@ public class WebDriverFactory {
     private final FixtureProviderService fixtureProviderService;
     private final ElementProxyCreator proxyCreator;
     private final BrowserStackRemoteDriverCapabilities browserStackRemoteDriverCapabilities;
+    private final SauceRemoteDriverCapabilities sauceRemoteDriverCapabilities;
 
     private final Integer EXTRA_TIME_TO_TAKE_SCREENSHOTS = 180;
 
@@ -122,6 +124,7 @@ public class WebDriverFactory {
         this.fixtureProviderService = fixtureProviderService;
         this.proxyCreator = proxyCreator;
         this.browserStackRemoteDriverCapabilities = new BrowserStackRemoteDriverCapabilities(environmentVariables);
+        this.sauceRemoteDriverCapabilities = new SauceRemoteDriverCapabilities(environmentVariables);
     }
 
     protected ProfilesIni getAllProfiles() {
@@ -150,11 +153,11 @@ public class WebDriverFactory {
     }
 
     public boolean usesSauceLabs() {
-        return StringUtils.isNotEmpty(ThucydidesSystemProperty.SAUCELABS_URL.from(environmentVariables));
+        return StringUtils.isNotEmpty(sauceRemoteDriverCapabilities.getUrl());
     }
     
     public boolean usesBrowserStack() {
-        return StringUtils.isNotEmpty(ThucydidesSystemProperty.BROWSERSTACK_URL.from(environmentVariables));
+        return StringUtils.isNotEmpty(browserStackRemoteDriverCapabilities.getUrl());
     }
     
     /**
@@ -279,7 +282,7 @@ public class WebDriverFactory {
     }
 
     private boolean saucelabsUrlIsDefined() {
-        return ThucydidesSystemProperty.SAUCELABS_URL.isDefinedIn(environmentVariables);
+        return StringUtils.isNotEmpty(sauceRemoteDriverCapabilities.getUrl());
     }
 
     private boolean browserStackUrlIsDefined() {
@@ -287,7 +290,7 @@ public class WebDriverFactory {
     }
     
     private WebDriver buildSaucelabsDriver() throws MalformedURLException {
-        String saucelabsUrl = ThucydidesSystemProperty.SAUCELABS_URL.from(environmentVariables);
+        String saucelabsUrl = sauceRemoteDriverCapabilities.getUrl();
         WebDriver driver = webdriverInstanceFactory.newRemoteDriver(new URL(saucelabsUrl), findSaucelabsCapabilities());
 
         if (isNotEmpty(ThucydidesSystemProperty.SAUCELABS_IMPLICIT_TIMEOUT.from(environmentVariables))) {
@@ -323,15 +326,7 @@ public class WebDriverFactory {
         String driver = getDriverFrom(environmentVariables);
         DesiredCapabilities capabilities = capabilitiesForDriver(driver);
 
-        configureBrowserVersion(capabilities);
-
-        configureTargetPlatform(capabilities);
-
-        configureTestName(capabilities);
-
-        capabilities.setJavascriptEnabled(true);
-
-        return capabilities;
+        return sauceRemoteDriverCapabilities.getCapabilities(capabilities);
     }
     
     
@@ -342,87 +337,6 @@ public class WebDriverFactory {
     	
         return browserStackRemoteDriverCapabilities.getCapabilities(capabilities);
 
-    }
-
-    private void configureBrowserVersion(DesiredCapabilities capabilities) {
-        String driverVersion = ThucydidesSystemProperty.SAUCELABS_DRIVER_VERSION.from(environmentVariables);
-        if (isNotEmpty(driverVersion)) {
-            capabilities.setCapability("version", driverVersion);
-        }
-    }
-
-    private void configureTargetPlatform(DesiredCapabilities capabilities) {
-        String platformValue = ThucydidesSystemProperty.SAUCELABS_TARGET_PLATFORM.from(environmentVariables);
-        if (isNotEmpty(platformValue)) {
-            capabilities.setCapability("platform", platformFrom(platformValue));
-        }
-        if (capabilities.getBrowserName().equals("safari"))
-        {
-            setAppropriateSaucelabsPlatformVersionForSafari(capabilities);
-        }
-    }
-
-    private void setAppropriateSaucelabsPlatformVersionForSafari(DesiredCapabilities capabilities)
-    {
-        if (ThucydidesSystemProperty.SAUCELABS_DRIVER_VERSION.from(environmentVariables).equalsIgnoreCase("mac"))
-        {
-            String browserVersion = ThucydidesSystemProperty.SAUCELABS_DRIVER_VERSION.from(environmentVariables);
-            if (browserVersion.equals("5"))
-            {
-                capabilities.setCapability("platform", "OS X 10.6");
-            }
-            else if (browserVersion.equals("6"))
-            {
-                capabilities.setCapability("platform", "OS X 10.8");
-            }
-            else if (browserVersion.equals("7"))
-            {
-                capabilities.setCapability("platform", "OS X 10.9");
-            }
-        }
-    }
-
-    private void configureTestName(DesiredCapabilities capabilities) {
-        String testName = ThucydidesSystemProperty.SAUCELABS_TEST_NAME.from(environmentVariables);
-        if (isNotEmpty(testName)) {
-            capabilities.setCapability("name", testName);
-        } else {
-            String guessedTestName = bestGuessOfTestName();
-            if (guessedTestName != null) {
-                capabilities.setCapability("name", bestGuessOfTestName());
-            }
-        }
-    }
-
-    
-    private String bestGuessOfTestName() {
-        for (StackTraceElement elt : Thread.currentThread().getStackTrace()) {
-            try {
-                Class callingClass = Class.forName(elt.getClassName());
-                Method callingMethod = callingClass.getMethod(elt.getMethodName());
-                if (isATestMethod(callingMethod)) {
-                    return NameConverter.humanize(elt.getMethodName());
-                } else if (isASetupMethod(callingMethod)) {
-                    return NameConverter.humanize(callingClass.getSimpleName());
-                }
-            } catch (ClassNotFoundException e) {
-            } catch (NoSuchMethodException e) {
-            }
-        }
-        return null;
-    }
-
-    private boolean isATestMethod(Method callingMethod) {
-        return callingMethod.getAnnotation(Test.class) != null;
-    }
-
-    private boolean isASetupMethod(Method callingMethod) {
-        return (callingMethod.getAnnotation(Before.class) != null)
-                || (callingMethod.getAnnotation(BeforeClass.class) != null);
-    }
-
-    private Platform platformFrom(String platformValue) {
-        return Platform.valueOf(platformValue.toUpperCase());
     }
 
     private Capabilities buildRemoteCapabilities() {
