+++ thucydides-core/src/main/java/net/thucydides/core/webdriver/capabilities/RemoteDriverCapabilities.java
@@ -2,8 +2,6 @@ package net.thucydides.core.webdriver.capabilities;
 
 import org.openqa.selenium.remote.DesiredCapabilities;
 
-import java.net.URL;
-
 public interface RemoteDriverCapabilities {
 
     String getUrl();
